Refactor CompositionPoly to make more dev-friendly (generic over PackedField only, not also ExtensionField)
Summary: Refactor change to CompositionPoly, making generic over only PackedField (and not also ExtensionField) This is more ergonomic and stable change.

Motivation: In small-field sumcheck, we eventually change our Multilinears to be defined over a large ExtensionField after doing the classical folding with a challenge from this large ExtField.

As written before this commit, the user of small-field sumcheck would be required to pass in two separate CompositionPoly definitions. This introduces a class of possible bugs because of the logic duplication.

Notes: Similar design decision changes to come in the near future.