Product Check Implementation
Summary: Allows a prover to prove the product check relation as defined in "Succinct Arguments over Towers of Binary Fields" paper

Notes: Slight naming mismatch between repo and the aforementioned paper. Specifically, what the paper refers to as the "merge" virtual poly, the repo refers to as "interleaved". All documentation shall use the repo's naming scheme. What the repo refers to as the "merge" virtual poly is similar to "interleaved", except that newly added selector variables in "merge" are the highest-indexed variables instead of the lowest.