Refactor to create MultilinearPoly trait
`binius::polynomial::MultilinearPoly<P>` is a concrete struct which represents a multilinear polynomial by a full slice of its evaluations over the hypercube. The issue is that in many cases we want to treat multilinear polynomials over a subfield as multilinears over an extension field. For example, a `MultilinearPoly<BinaryField8b>` and `MultilinearPoly<BinaryField128b>` can both behave as multilinears over the 128-bit field, in a sense. We can't just directly convert the former to the latter though, because then we would need 16 bytes to represent each field element instead of 1 byte.

This change renames the `MultilinearPoly` struct to `MultilinearExtension`, and creates a new trait for representing multilinear polynomials over a field, called `MultilinearPoly`. `MultilinearPoly` is implemented by any `MultilinearExtension<F>` where `F` is a subfield. The trait is designed so that trait objects can be created.