Complete SNARK example for bitwise-AND relation
This creates an example to run a bespoke SNARK for the bitwise-AND relation on committed vectors. By bespoke, I mean that we intend to add more wrappers for this functionality as the codebase progresses. This does not yet make use of the new batch PCS scheme. The evalcheck verifier will need to be updated.