Add batching support to evalcheck; adapt bitwise_and_proof example.
The core idea is a new structure `BatchCommittedEvalClaims` which is passed to evalcheck proving and verifying routines as a third mutable parameter. The rationale for that is that batched commitments can potentially span across higher level predicate boundaries (i.e. it's not unreasonable to batch all F2 vectors of same size together). When the batched claim can be checked with PCS directly, it can be obtained via `get_same_query_pcs_claim`, otherwise it may be expunged with `take_claims` and resolved via an additional sumcheck round (out of scope of this MR).