Add binary_tower_level to MultivariatePolyOracle
High-Level reason for this change: When we need to mix polynomials for batching zerocheck or sumcheck, we would like to do so in the most efficient way and with minimal soundness degradation to the resulting protocol. Tracking the binary tower level of a MultivariatePolyOracle helps us do this as well as we possibly can.