[evalcheck] Fix bug with nested composite proofs
There was previously a bug with proving evalcheck when the claim was a multivariate polynomial defined over a linear combination multilinear. The problem was that multiple layers of the recursive evalcheck proof needed inner evaluation hints. This makes the proving code more robust by having a single witness type that maps oracles to MultilinearPolys which can be used to evaluate them.