[oracle] Introduce MultilinearOracleSet for consistent oracle IDs
This change introduces a new mechanism for constructing and accessing multilinear polynomial oracles. MultilinearPolyOracle is a self-contained recursive data structure, where each oracle is a tree of derivations over leaf oracles. Consistent indexing can reduce complexity in evalcheck and mixing of multilinear composites, and help with deduplication of oracles and associated witnesses.