Introduce MultilinearWitnessIndex to aid evalcheck
The primary change is introducing `MultilinearWitnessIndex` structure to index multilinear witnesses by their unique `OracleId`. This is populated by the highest level of the SNARK prover. It is used primarily in `evalcheck` to evaluate multilinears that are used in linear combinations and composites, and eventually will be used to construct witnesses for subsequent sumcheck rounds.