[poly_commit] Eliminate TensorPCS limitation causing huge proof sizes
This fixes an issue in the TensorPCS that lower bounds the number of rows of the matrix to the packing width, preventing optimal sizing. This limitation was there to accommodate the scalar transpose in the verification routine. Normally it's not a problem, unless there are many multilinears committed in a single batch, as is the case with some new examples.

We get around this limitation by padding before doing the scalar transpose, then trimming off the padding coefficients before verifying column consistency.