[sumcheck] Process batches of vertices when evaluating subcube in sumcheck
It appeared that the most of the time in sumcheck is spent on evaluating subcubes. So to make it in the most efficient way we're now passing a batch of vertices to evaluate them withing one function. This allows to make negligible the cost of virtual function call and bounds checking inside. Also the loop in `evaluate_subscubes` is that hot that using unsafe access inside also saves several seconds.