Witness index for native multilinear extensions
This MR introduces a new witness index that can return `MultilinearExtension` witnesses for heterogeneous fields. This is a potential solution to an issue in the example SNARKs where we must have a trace witness struct in addition to the trace oracle and the witness index. The witness index seems like it should be enough but is insufficient because the PCS needs direct access to `MultilinearExtension`, not the type-erased form of `MultilinearPoly` .

This new index exploits the fact that Underlier types can be packed into different extensions and a small amount of unsafe lifetime transmutation to provide a witness that can return native witness polynomials for a variety of fields.