[ntt] Refactor additive NTTs with on-the-fly and precompute into a single struct
This reduces code duplication and paves the way for a cleaner multithreaded NTT algorithm implementation. There is a single generic struct SingleThreadedNTT that is generic over the twiddle access method now.

This introduces a Builder-like pattern for constructing NTTs. For example, instead of constructing an NTT with precomputed twiddles directly, it is recommended to call `SingleThreadedNTT::new(...)?.precompute_twiddles()`.