[sumcheck] Switchover round heuristic
Sumcheck switchover rounds are hardcoded in the examples. Write a function the heuristically estimates the switchover round from the extension degree. The simplest heuristic is to pick the switchover round to be the first round when the stored multilinear is 2^k sized smaller _in bytes_ than the original multilinear, for some constant k. That is, if the multilinear coefficients are a 1/2^d sized subfield of the big field, then after round (d + k), the folded multilinear will be the size of 1/2^(d+k+1) big field elements.