[protocols] Abstract sumcheck refactor to make batching the default strategy
The grand idea of this refactor is to bridge the gap between batched and singleton abstract sumcheck instances - because all sources of overhead in the former are totally avoidable and are simply the result of choosing unfitting abstractions. Apart from performance gains associated with sumcheck batching, this should provide uniformity (only batched sumchecks are left to be handled) and better separation of concerns - there is common state and individual state in a batch, which removes eligibility checks, etc.

High-level description of the approach:

1. `AbstractSumcheckClaim` and `AbstractSumcheckWitness` become traits - anything that is able to provide oracle, composition and multilinears fits
2. `AbstractSumcheckProversState` trait is introduced - which stores common state for batched abstract sumcheck instances, updates it prior to executing batched rounds, and can spawn individual provers from claim/witness pairs - _this removes performance overhead associated with batching_
3. Change batching interfaces to take witnesses directly instead of premade prover states - _this is a usability improvement_