[challenger] Refactor to deduplicate HashChallenger and DuplexChallenger
This introduces a new FieldChallenger struct that can be generically instantiated with either a hasher or a duplex sponge. This deduplicates the logic that we have for HashChallenger and DuplexSpongeChallenger.

This MR only introduces the new challenger type, it does not remove the old code nor upgrade the code that uses it. That should be done as follow-on merge requests.