[fri] Augment FRI protocol to handle batched FRI for interleaved codes
There is an optimization available for FRI, which is compatible with the FRI-Binius multilinear PCS, that instead of encoding with and testing proximity to a Reed-Solomon code, the protocol can encode with and test proximity to an interleaved Reed-Solomon code. This saves a couple of NTT butterfly rounds, but more importantly reduces the total amount of Merkle hashing by the batching factor. This comes at no loss (and even a slight improvement) in proof size as long as the log batch size equals the first round folding arity that FRI would have used anyway.