[FRI] Optimize FRI final round
In the FRI protocol query phase, we currently run the same procedure for all test queries: we go through each pair of consecutive folded codeword commitments and test consistency of a coset of the earlier commitment with a point in the later commitment. The FRI protocol folding phase ends when the prover folds down to a RS codeword with dimension 1. We know that a RS codeword with dimension 1 is a repetition code; that is, all of the codeword symbols are equal and in fact they are equal to the sole message symbol.

This is somewhat wasteful because we observe that in the folding of the last committed message down to a single symbol of the dimension-1 codeword, there are very few cosets we can fold, yet we do it redundantly for each of the test queries. Typically we will have 100-200 test queries and a final codeword length of only 2 or 4.

Instead now, the prover sends the entire decommitted codeword for the last commitment to the verifier, who folds it in its entirely and checks that it is a dimension-1 RS codeword.