[zerocheck] Univariate skip evaluation for zerocheck
Given that univariate skip is a single round, its evaluation is implemented as a mere function `zerocheck_univariate_evals`, which takes in a zerocheck witness (multilinears & compositions) and zerocheck challenges and computes the univariatized round polynomial over first `skip_rounds` variables. It is assumed that sampling the challenge, obtaining multilinear partial evaluations, and bootstrapping regular zerocheck provers for the remaining rounds would be the responsibility of a batching protocol adapter.