[sumcheck_v2] Refactor prover_state to extract round eval computation to new module
The goal of this refactor is to make `ProverState::calculate_round_evals` standalone, not relying on the `&self` parameter, so that it may be extracted to the `binius_hal` crate. As an intermediate step, this moves the computation to a new standalone module `binius_core::protocols::sumcheck_v2::prove::round_calculator`.