[piop] Full FRI batching in polynomial IOP compiler
This creates a new `binius_core::piop` module that is intended to replace `poly_commit`. It implements the multilinear sumcheck IOP compiler using FRI-Binius. That means it exposes a commit-prove-verify interface allowing the caller to commit a collection of multilinear polynomials together, then prove and verify a group of sumcheck claims with a special form. The form is that it proves sumchecks that are bivariate products of a committed polynomial and a transparent polynomial.