[ring_switch] Standalone ring-switch reduction
New implementation of the ring-switching reduction that does not depend on the deprecated PolyCommitScheme trait. This module is intended as a reduction to the PIOP compiler in the `piop` module. This implements a fully-batched version of the ring-switching protocol that handles many polynomials of different tower heights and numbers of variables.