[zerocheck] Zerocheck base field detection
Invocations of `constraint_system::prove` do not need to specify `FBase` anymore, as it is determined automatically for each zerocheck prover as the smallest tower field that can embed all multilinears and composition constants, if any. For some circuits this may result in performance improvements due to `FBase` previously being global and thus overly conservative.

This has been achieved via:

1. boxed receiver of the `finish` method in `SumcheckProver` to make that trait object safe (it previously wasn't despite claim in the comments)
2. boxed receiver of the `fold_univariate_round` and type erased returned prover in `UnivariateZerocheckProver` to also make it object safe
3. `constraint_system::prove` operates over type erased provers