Improve test compilation time (#10)
Co-authored-by: Dmytro Gordon <dgordon@irreducible.io>

This MR addresses tow issues that make cargo test slow:

Thin LTO slows down compilation of all the crates a bit.
It takes quite a time to compile and link all the examples with test profile which are not actually executed. So I've added an alias to compile and run tests only for fast local usage.