[m3] M3 gadgets for Grøstl-256 permutations (#139)
This re-introduces Grøstl M3 gadgets using the new crate. This supports both P and Q permutations and tests against the portable native implementation. The gadget operates over an isomorphic field representation in the canonical tower basis, because the constraints can be expressed more efficiently in that manner.

This also re-introduces the end-to-end Grøstl example program.