[sumcheck] Introduce a struct for a batch of rows of the same length (#161)
It appeared that we are creating that big batch of values when proving Vision32 that just checking if all the rows in the batch are of the same length takes signifficant amount of time since we are doing it for each batch_evaluate call. in this PR we introduce a special struct that guarantees that all the rows are of the same length by checking it at the moment of creation. Also I like this approach more because this struct helps self-describe the argument of batch_evaluate.