[m3] Refactor away duplication of table size specification (#211)
Previously all table sizes needed to be specified when the witness index was created. There's enough information in the fill_table_sequential and fill_table_parallel calls to initialize the table indexes properly, so this change allows the table initialization to use the data from those calls. When creating the Statement for constraint system compilation, it's recommended to get the table information from the witness, or in the verifier's case, received from the prover.