Deduplicate events when tracing multiplications (#203)
Before we used to add an event for every mutliplication function call, which resulted in a certain level of duplication in case when the function redirects to another one or calls the multiplicaiton recursively for a smaller level. E.g. if we were multiplying two 128b scalars we generated several events:

    Scalar field multiplicaiton128b x128b
    Packed field multiplication 1x128b x 1x128b
    Two packed field multiplications 2x64b x 2x64b
    ...

In this PR I'm adding a thread-local boolean flag and a guard that set's it until the multiplicaiton function is over. If this flag is set we do not log new events. Here we depend on the fact that we never split a single multiplication between several threads.