[fiat_shamir] Update challenger to use initial digest (#218)
This updates the challenger so that it's pre-seeded with a Grøstl digest digest; when it starts observing for the first time, its observations will be appended to digest. right now, the first batch of observations are appended to [], while future batches get appended to some internal digest (itself gotten as a hash of the last round of observations).

in the future, we will want said initial digest to be a hash of the constraint system / statement, etc. Right now, it's just Grøstl("").

Co-authored-by: Joseph <jjohnston@irreducible.com>