[piop] Handle arbitrarily small committed polynomials (#227)
This relaxes the ring_switch and piop modules to handle committed polynomials that are smaller than the size of a packed large element and eliminates the OracleTooSmall error. We handle this case by committing the polynomial as it's MLE over the larger hypercube, which corresponds to repeating it's evaluations. The integration test in ring_switch module is updated to exercise this code path.