SYS-110: compute API supports host buffer management, allocation
Accelerated compute backends typically require setting up IOMMU or
performing custom allocation and alignment for buffers that are copyable
to and from device memory. Extend the `ComputeMemory` trait with a
type that is transparent to Binius as a dense slice of elements with a
minimum slice width.

Define a `ComputeBufferAllocator` trait supporting
allocating host and device memory. Implementations will provide
specializations of `ComputeMemory` and `ComputeBufferAllocator` that
perform the device-specific setup for both classes of allocations in
addition to specifying the implementation-specific constraints on
minimum span widths.