[m3] Fix the small-table flushing bug for lookups (#230)
This issue is caused by the solution of #170, which makes it possible for tables to be automatically allocated capacity exceeding their event count. That works fine for regular tables, but not lookup tables that are power_of_two_sized. However, that solution is now obsolete because of an alternative fix in #227, so this adds a test and assertion and rolls the change back.