[m3] High-level model for Grøstl-256 hash function
Implementation of the high-level model of Grøstl-256. The presented version
is specialized to a fixed length input size: 1024 bytes. This is done
because otherwise it would be harder to arithmetize the logic of adding
padding. It would be easy to change the length of the input as long as it is
a multiple of Grøstl-256 block size (64 byte).

See [CRY-312].

[CRY-312]: https://linear.app/irreducible/issue/CRY-312/high-level-model-for-grostl-256-hash-function

Co-authored-by: Jim Posen <jimpo@irreducible.com>