simpler approach
see code for comments / remarks. the code turns out to be simpler to write if we allow `log_batch_size` to stay ≤ arity, as opposed to letting it get as large as `total_vars`. this only affects the case of empty oracles. the point is that when oracles are empty, we COULD just go full interleaved---but doing that makes the code more complicated to write and read, for various ultimately artificial reasons, but nonetheless.