[ntt] Add skip_rounds parameter to AdditiveNTT (#306)
This generalizes the AdditiveNTT operation to be able to perform a truncated NTT, which skips rounds at the beginning of a forward transform and the end of an inverse transform. This is useful for Reed-Solomon encoding, where we want to run a large NTT but know apriori that the first few rounds have the effect of simply copying data.

This resolves a limitation of ReedSolomonCode::encode, which imposes a packing length restriction on the encoded message and also complicates the multithreading around the NTT with an additional Rayon loop. Now the multithreading is delegated to the AdditiveNTT implementation, which can size the matrix more appropriately.