[core] newtype OracleId
OracleId becomes a newtype. This is a non-functional change.

As you see, this PR introduces some friction when using OracleId as an index by
forcing using `OracleId::from_index` and `OracleId::index`. This is intentional.
Users who call either of those functions should think of a better way of
structuring the code to avoid this. Some of the usages are hopefully going away
with follow-up refactorings.

As a bonus: turns out there were quite a bit of clones of MultilinearPolyOracle
struct. Each one is, if instantiated with e.g. F=BinaryField32b is 160 bytes.
Not too much but definitely annoying, especially considering why it was
happening. In one instance the whole clone of the structure would be done
just to get the name/label.

The tests seem to pass locally and petravm also green.