[m3] High-level model for Grøstl-256 hash function (#310)
Implementation of the high-level model of Grøstl-256. The presented version is specialized to a fixed length input size: 1024 bytes. This is done because otherwise it would be harder to arithmetize the logic of adding padding. It would be easy to change the length of the input as long as it is a multiple of Grøstl-256 block size (64 byte).