[core] newtype OracleId (#327)
OracleId becomes a newtype. This is a non-functional change.

As a bonus: turns out there were quite a bit of clones of MultilinearPolyOracle
struct. Each one is, if instantiated with e.g. F=BinaryField32b is 160 bytes.
Not too much but definitely annoying, especially considering why it was
happening. In one instance the whole clone of the structure would be done
just to get the name/label.