[m3] Indexed lookups with tallying (#331)
Indexed lookup tables are fixed-size tables where every entry is easily determined by its index.

Indexed lookup tables cover a large and useful class of tables, such as lookup tables for bitwise operations, addition of small integer values, multiplication, etc. The entry encodes an input and an output, where the index encodes the input. For example, a bitwise AND table would have 2 8-bit input values and one 8-bit output value. The index encodes the input by concatenating the 8-bit inputs into a 16-bit unsigned integer.

Before a lookup table witness can be filled, the number of times each entry is read must be known. Reads from indexed lookup tables are a special case where the counts are difficult to track during emulation, because the use of the lookup tables is an arithmetization detail. For example, emulation of the system model should not need to know whether integer additions within a table are constraint using zero constraints or a lookup table for the limbs. In most cases of practical interest, the lookup table is indexed.

The method to tally counts is to scan all tables in the constraint system and boundaries values, and identify those that pull from the lookup table's channel. Then we iterate over the values read from the table and count all the indices.