[utils] introduce mem::slice_uninit_mut (#355)
As per @GraDKh request in this [comment].

I did not use the suggested identifier as, unlike in `assume_init`, we don't
really assume anything here, we are just converting a definitely inited slice
into a maybe uninited slice. Also to make it inline with the follow-up I
decided to call it `slice_uninit_mut` (put slice in the front).

[comment]: https://app.graphite.dev/github/pr/IrreducibleOSS/binius/353/%5Bcore%5D-miri