[m3] don't assume oracle id order in witness (#357)
This changeset aims to fix the implicit dependency of the witness on the order
of oracles added in `ConstraintSystem::compile`.. See [CRY-356] for more
details.

This does several things:

1. Prefers using ColumnId over indices
2. Introduction of `OracleLookup`

This starts using the ColumnId where possible. This change should really make
much less likely to confuse indices and improve general type-safety, and as well
as simplify the code a bit (one type is better than multiple). However, this
makes the column IDs to carry table_id around, but I don't think this is a
problem since the code is not hot.

Introduction the `OracleLookup` type. Prior this changeset
the data about which oracle came from which columns was lost. This type
is meant to store this information explicitly to be read by the witness
handling code.

This opens up for other use-cases where you might want to pass some data
from the compilation process into the witness generation, as that seems to be
inherently important: compilation creates oracles from columns and witness
filling operates on oracles.

Ideally though IMO we should really introduce a special type that represents
the compilation output:

1. `TableStat` should be a function of the compiled constraint system. That
   said, we can easily account for that.
2. Warnings. Optionally catch silly things like a committed column that is not
   referenced in a virtual column or a constraint.
3. It will also contain the data similiar to OracleLookup.

[CRY-356]: https://linear.app/irreducible/issue/CRY-356/dont-assume-the-order-of-compilation-of-oracles