[ntt] Add coset_bits parameter to AdditiveNTT (#371)
Modify the AdditiveNTT behavior so that the evaluation domains for differently-sized transforms are FRI-compatible. By FRI-compatible, we mean that encoding then FRI folding commutes with tensor folding then encoding. This is checked with a proptest in the `fri` module.

The interfaces for AdditiveNTT should be cleaned up to use structs with named fields and Default impls instead of many parameters. I'm leaving this for future work.

Fixes CRY-71