Environment variable flag to dump transcript to a file (#367)
### TL;DR

Added functionality to dump proof transcripts to a file when the `BINIUS_DUMP_PROOF` environment variable is set.

### What changed?

- Modified the `finalize` method in `ProverTranscript` to check for the `BINIUS_DUMP_PROOF` environment variable
- When the environment variable is set, the proof transcript is written to the file specified by the variable's value
- Added necessary imports for file operations (`std::fs::File`, `io::Write`)
- Made a minor refactoring in `into_verifier` method to store the transcript in a variable before passing it to the constructor

### How to test?

1. Set the environment variable `BINIUS_DUMP_PROOF` to a file path:
   ```
   export BINIUS_DUMP_PROOF=/path/to/output/proof.bin
   ```
2. Run a process that generates a proof
3. Verify that the proof has been written to the specified file

### Why make this change?

This change enables easier debugging and analysis of proof transcripts by allowing them to be dumped to a file for inspection. This is particularly useful during development and testing of cryptographic protocols, as it allows for offline examination of the proof data.