[compute] Introduce ComputeLayer::extrapolate_line operation (#391)
# Introduce ComputeLayer::extrapolate_line operation

This PR adds a new `extrapolate_line` operation to the `ComputeLayer` trait. This operation extrapolates a line between evaluations at 0 and evaluations at 1, computing `y_z = y_0 + (y_1 - y_0) * z` for a given point `z`. The operation is performed in parallel over vectors of field elements.

The implementation includes:
- Adding the `extrapolate_line` method to the `ComputeLayer` trait
- Implementing the method for `CpuLayer<T>`
- Adding test utilities and a test case to verify correctness
- Making `ComputeLayer<F: Field>` and `TowerFamily` require `'static`

The operation is useful for interpolating between two points on a line and evaluating at arbitrary positions.