[core] miri (#353)
This changeset is related to [CRY-341]. Specifically it patches some cheesy
spots but it also highlights the issue using miri.

To run the test you need to execute the command below:

    RUSTFLAGS="-C target-feature=-neon" cargo miri test -p binius_core -- batch_evaluate_add_mul

Disabling -neon is required for apple silicon hardware as the code tries to use
the SIMD optimized versions of field math. For intel hardware you might need
to do a similiar tweak and is left as an exercise for the reader.

Besides that, the stackalloc was also wrapped as `alloca` is not recognized
by miri.

[CRY-341]: https://linear.app/irreducible/issue/CRY-341/check-funny-business-in-arith-circuit