[core] introducing structured oracle type
Structured is an oracle type that is very similar to Transparent, but is
defined by an ArithCircuit of some upper bound of n_vars.

This oracle type is mainly motivated by the needs of structured columns. The
idea is that such a structured column will define an arith expression up to some
maximum size. The frontend can supply any table size up to that maximum.

See [CRY-396] for more details.

[CRY-396]: https://linear.app/irreducible/issue/CRY-396/transparent-oracle-type-based-on-arithcircuit