[Evalcheck] remove redundant query memoization (#452)
It appears that the performance boost is not primarily due to storing partial_eval, but rather due to splitting the computation into two parts. When replacing the suffix+prefix split with a simple mid split, performance only slightly degrades.

My hypothesis is that this behavior is related to cache usage. For example, the tensor product of a point with 25 variables takes around 500MB, whereas with only half the variables it's approximately 0.13MB — a dramatic difference that likely affects memory locality and cache efficiency.

In any case, this PR improves performance from 2x (Groestl) up to 10x (SHA-256) compared to the previous implementation.

It's also quite possible that similar optimizations could be applied to other parts of the codebase in the future.