SYS-258: fix `test_generic_fri_fold` host buffer (#491)
* SYS-258: fix `test_generic_fri_fold` host buffer

The prior version invoked `copy_h2d` with a host buffer that was not
allocated with the host allocator. Using a buffer allocated with the
default allocator results in incorrect behavior if the hardware
implementation derives additional addressing information during copies
from the provided slice assuming that the slice was allocated by the
implementation's host allocator.

* Fix clippy