[core] remove fn oracle (#475)
As you can see in the changeset, this function was overused, which lead to
copying the entire oracle data which could be, IIRC, up to 180 bytes. Hence we
just remove that function. In case, the actual copy is required `Clone::clone`
is suggested, or better, surgical copy of only data required.