Remove unnecessary optimization passes in `determine_const_eval_suffixes` (#527)
In `determine_const_eval_suffixes` we are calling `optimize` for the expressions where we only need to get a constant value from. This involves deduplicating steps operation which becomes quite slow in case when there are many expressions to process.

What we need is to only perform a constant folding, so in this PR this method is exposed to be public.